First fix for bringing the popc version up to date. Some warnings in IOUtils and Meteo2DInterpolators have been fixed (non explicit cast).

+++ src/IOUtils.h
@@ -27,8 +27,8 @@ using namespace std;
 namespace IOUtils {
 
 	const double nodata = -999.0;	//HACK: we should define the same nodata everywhere...
-	const unsigned int unodata = -1;
-	const unsigned int npos    = -1;
+	const unsigned int unodata = (unsigned int)-1;
+	const unsigned int npos    = (unsigned int)-1;
 
 	/**
 	* @brief Coordinate conversion: from WGS84 Lat/Long to Swiss grid
